llmpackagequery/requirements.txt
openai
gpt4all[cuda]
ollama
google-genai
mistralai